Switched to a snapshot-based system.
Instead of constantly hitting the pipe, we grab them once and cache the results
in registers (on x86_64) or the stack due to register pressure (on x86). This
keeps us from fucking up our data cache with a whole bunch of shared data
accesses, and makes sure each thread has a single, coherent view of the pipe.

Now, every `static inline' function is inlined, and speed is through the roof.